+++ src/gtkhx.c
@@ -1269,14 +1269,14 @@ on_user_info_signal (GtkhxSession *emitter, guint uid, gpointer nam,
 static void
 on_file_info_signal (GtkhxSession *emitter, gpointer path, gpointer name,
                      gpointer creator, gpointer type, gpointer comments,
-                     gpointer modified, gpointer created, guint size,
+                     gpointer modified, gpointer created, guint64 size,
                      gpointer user_data)
 {
     (void)emitter;
     (void)user_data;
     output_file_info ((char *)path, (char *)name, (char *)creator, (char *)type,
                       (char *)comments, (char *)modified, (char *)created,
-                      (guint32)size);
+                      size);
 }
 
 static void
